+++ purchase-service/src/main/java/com/shop/purchaseservice/DTO/ProductDuplicateDTO.java
@@ -22,6 +22,6 @@ public class ProductDuplicateDTO {
     private String producer;
     private String category;
 
-    private Double feedBack;
+    private BigDecimal feedBack;
 
 }
